[media][tests][e2e] Synchronization callbacks for virtual audio
This CL and two upcoming ones comprise a significant set of
changes to the virtual audio FIDL interface and driver.

In this CL, callbacks are added to the Enable and Disable methods,
to allow for clients to synchronize with the full removal of
devices.

Also included are the reformatting of the virtual_audio.fidl
and audio_capturer.fidl files, and the renaming of virtualaudio's
ResetConfig to ResetConfiguration (per API Review feedback).

Test: build, CQ

Change-Id: I97355a49827c7ba43ef9423b97e942e64525cae5