[zircon][block-watcher] Add the ability to disable the block watcher in an isolated devmgr
This patch adds the capability to disable the block watcher from within an isolated devmgr.
This is necessary to work around a deadlock that can occur if the watcher calls read
while a block device is being unplugged. This deadlock would cause the devhost
to hang indefinitely (or until the watchdog timer killed it).

Test: runtests -t usb-virtual-bus-test
The above test will be integrated in a future CL

Change-Id: Iae4e994a7f69323e2702b362f60504b68c58a094