[media][drm] Update ContentDecryption API
While there are a commonly used initialization data formats, content
decryption modules may define or support additional formats beyond what
we have listed in our API. Per the API readability rubric guidance, this
unbounded set of formats should be treated as a string rather than an
enum.

Also per the readability rubric, DecryptorParams has been updated to be
a table rather than a struct, as it is likely to change in the future
and gain additional fields.

This change should not impact SDK users as this library is not yet in
use.

Change-Id: I635c2120663f8312ee12b4abcb37cc51664cd16a