[zircon][usb-virtual-bus] Move the USB virtual bus into its own devmgr
This patch moves the usb-virtual-bus into its own devmgr, and adds a test binding for it.
It also removes usb-virtual-bus functionality from usbctl, as this will be tested in an
automated fashion in a future CL.

In order to control the virtual bus, it is necessary to start an isolated devmgr
 and create a bindable device with BIND_PLATFORM_DEV_VID set to PDEV_VID_TEST and
BIND_PLATFORM_DEV_PID set to PDEV_PID_USB_VBUS_TEST.
Then; ioctls can be used to manipulate the virtual bus.

Test: Validate that usbctl no longer contains any virtual bus functionality, and that the virtual
bus is no longer visible in the device tree.

Change-Id: Ia3ae6eb89c71ecafae95b427b1cb0659da206fb5