[health_checker] Attempt reboot if unhealthy.
If the state is found to be unhealthy, it will still be returned as unhealthy.
However, if the reboot flag is set to true, then the health_checker will attempt
to reboot the device as well without checking the state again after the attempt.
If the device has been rebooted successfully, it should show up as healthy on the
next call to the health_checker tool.

Bug: IN-791 #comment
Change-Id: I32ad73398ccd1819b832519f49b5e3c4313a26e8