[mediaplayer] don't DCHECK on unnecessary packet request
The ffmpeg demux node in mediaplayer occasionally gets a request
(from downstream) for a packet even when it has reported
end-of-stream on all streams. This happens only rarely.

https://fuchsia-review.googlesource.com/c/fuchsia/+/263095 addressed
the first-encountered DCHECK. That just got us to the next DCHECK.
This CL should prevent this failure in CQ going forward.

MTWN-242 #done

TEST: fx shell run_test_component mediaplayer_tests \
      --gtest_repeat=<many>
      # debug build
Change-Id: I0b070760621d0f8b78124be44aa28a5358b3c2a7