[upci] Preserve existing BAR allocations
Some PCI devices like serial consoles and framebuffers are
brought online by the BIOS/bootloader and have existing BAR
allocations. For BAR bringup this was not maintained in the
userspace driver, but for continued work moving forward it
was necessary.

If the bus driver sees something that could be a valid address
in the address bits of a given device BAR then it will attempt
to allocate that address from the upstream allocator, if and only
if the device has been enabled prior to our probing. Unfortunately,
at power-on the state of all address bits are indeterminate, so
the best we can do is preserve any address we see present and then
attempt to use it. If we can allocate it then we don't need to change
anything and we can proceed. If not, then we allocate a properly
sized region from upstream and continue with BAR configuration.

This should match the existing behavior in the kernel PCI bus driver.

ZX-3146

Change-Id: I974d903b9d2a4e99069b874a1a2c0389c9e50630