[ulib] Initial check-in of Scudo standalone
The standalone version of Scudo is currently being upstreamed
in LLVM's compiler-rt, this is meant to provide a temporary way
for Fuchsia to use it as a replacement for jemalloc in musl.

Some of the files have already been reviewed and commited upstream,
while others haven't, and are likely to undergo some changes
during the review process.

The LLVM copyright header is present in all files, even the ones
that are not yet upstream, which I am not sure is legitimate.

BUILD.gn was provided by David. Note that hardware CRC32 is not
enabled.

Test: replace jemalloc with scudo in musl, fx full-build & fx run.
Change-Id: I42d6dbefd25e7c2252a18f712243b72b73b04ed3