[minfs] Lock during transaction scope
When a Transaction is started, acquire a lock from
the Minfs instance. Currently all Minfs operations
are performed synchronously, but this will prevent
future asynchronous operations from attempting to
modify metadata structures simultaneously.

Specifically, once data blocks are allocated
asynchronously, the super block, block bitmap, and
inode table may be accessed at the same time.

Add Transaction as a requirement for more methods
to ensure that the lock has been acquired first.

Add an additional WritebackWork to the Transaction
to be used for data block updates, as well as more
methods for interacting with the block and inode
promises.

Test: fs-test

Change-Id: I3c075306d468b105319faad9e993f81bf835385c