[banjo][ispimpl] Workaround banjo heuristics issue.
Banjo v1 doesn't have proper type informtion to determine when to to
expansion it make heuristics based decisions based on keywords. cb
triggers one of the keywords, so changing the member name is a
workaround necessary to allow us to get correct behavior which lines up
with what banjo v2 will do (regardless of the method member name).

Tested: Compiled
Change-Id: I7ee083189f400090b31c4b211597e34586416fd5