[system-ota-tests] Fix races when starting up repo serving
There were two issues with serving repositories. First, since we
were spawning the repo serving on a goroutine, we were being racy
between starting the server and starting the tests. This changes
the server to block until we are listening before starting the
tests.

In addition, we were also competing with other "pm serves" on trying
to use port 8083. Instead, we'll just have the system pick us a port
to listen.

Test: Verified ota tests still work
Change-Id: I479c81c04c11810390589b5ba5a840b9dd46a9b5