[debian_guest] Work around an issue with bootstrapping Debian
Debian contains two packages, base-passwd and base-files, which have an
implicit dependency. Debootstrap knows this and handles it by
ensuring that base-passwd is installed first and this CL ensures that
our multistrap configuration script does the same.

See this Debian bug report for an extended discussion on the issue:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=924401

Also add procps to the package list mainly for ps, which is useful for
debugging.

Change-Id: I4f555209822e4a781ad4f54d5299ad4b481256fc