[media][e2e] Cleanup VAD memory ownership
This CL addresses memory ownership and lifetime, in both control and
stream devices. Regarding the former, VirtualAudioControlImpl now
calls remove_device when its Unbind method is invoked.

The latter is more interesting. There are five different cases that
trigger a stream shutdown:

1) Control device is unbound (or more generally the singleton's
   ~VirtualAudioControlImpl is invoked for any reason, but note that
   the DDK references this object even during DdkUnbind until it calls
   DdkRelease) -- triggered by DevHost or some other manipulation of
   the device tree.

2) Stream Input/Output FIDL channel disconnects
   (or more generally, the parent ~VirtualAudioDeviceImpl occurs, but
   note that the FIDL binding holds this object's unique_ptr) --
   triggered by the FIDL client: either an intentional channel reset,
   or an error that forces a channel-disconnect.

3) Control::Disable FIDL method is called
   (this should destroy all active stream devices, while retaining
   their FIDL channels and states) -- called by FIDL client.

4) Input|Output::Remove FIDL method is called
   (this should destroy that one stream device if active, while
   retaining its FIDL channel and state) -- called by FIDL client.

5) Stream device is unbound
   (DdkUnbind subsequently leads to DdkRelease and ~VirtualAudioStream
   but should leave the FIDL channel and update the DeviceImpl to be
   aware that the stream has departed) -- triggered by DevHost or some
   other manipulation of the device tree.

For the first four cases, RemoveStream marks a flag to signal that the
stream-deletion is originating "from above", then calls the stream's
DdkUnbind. During DdkUnbind, if stream-deletion is NOT originating
"from above" (bool shutdown_by_parent_), the stream instructs the
parent to "forget about us" so that it is not later double-freed
(also, in cases 1 & 2, we are in ~VirtualAudioDeviceImpl during the
stream's unbind, so it cannot interact with the device in any way).

Test: Build, CQ, manual testing using virtual_audio CLI utility.

Change-Id: I6b8fcb3a13ee6b63c4aee6a664d94fd928637b8b