[build] Move Zircon GN files to the top
This moves the generation of the Zircon GN files to the top of the file,
so that they are generated before anything else gets queued.

This change is because some of the buildid transition ended up tickling
the inherent race condition here, in that we depend on the fact that
this generation completes before any target attempts to load a Zircon
dependency.

The combination of recent additions to the top of this file and an
attempt to add a set of default dependencies to support buildids
triggered a file-not-found error, as it rearranged the file loading
order a bit and the Zircon files were not finished soon enough.

This change should allow the Zircon files to finish prior to
triggering any other file loads.

Bug: TC-399 #comment
Change-Id: I23196bf666e4f9c2b346723ec9e27004f576bbc5