[bt][tests] Add async expectations and LE types to fuchsia-bluetooth lib
Add the `expectation::asynchronous` module to fuchsia-bluetooth, which
allows tests to be written with expectations that are expected to be
fulfilled asynchronously at some point in the future.

Also add the `le` module which redeclares simplified types of the FIDL
data structures for LE, collapsing optionals and implementing `Clone`
for ease of use. This allows FIDL types to be converted into these types
at program boundaries, and then simplifies use internal to the program.

Change-Id: Ic6933dffff32b7324436ba6d72471325056b3a3c