[pkgfs] dynamic index caches active blobs
This is not an optimal fix for the problem. We will want to invert the index
to just be an index of blobid's in the future, but it also should be hoisted
up out of pkgfs.

Instead pkgfs should support opening blobs by merkleroot, but should validate
the package validity and blob needs at open time.

Bug: PKG-588 #done
Bug: PKG-486 #done
Change-Id: I15a709e7632812a4827fddae9b33d554078b3552