[debugger] Add more tokens.
Type parsing needs "&&" for rvalue references so this token is added.

For testing purposes, the "&" token as a binary operator is added. And
for symmetry | and || are added.

Tokens are added for keywords required for type parsing are added:
"const", "volatile", "and restrict".

None of these new operators are currently hooked up to any operations.

Change-Id: Ie8784376d551f5c5108e7742f5ef9e8feba4a6bf