[minfs] Add 'swap' block operation
Previously for allocated blocks, the 'write'
operation would either return the block (for
directory vnodes), or swap the existing block
with a new block (for data vnodes). This patch
splits the swapping functionality into a new
'swap' operation.

Additionally, the relative index of the block
being op'd is now tracked through the BlockOp
stack.

Test: fs-test

Change-Id: I6b9a4290bd6180ffde4ae412856b0ca8220a83f8