[minfs] Perform pending block swaps in AllocateData
Instead of swapping blocks in the middle of write
and truncate operations, mark blocks for allocation
and then swap them all at once. This requires the
VnodeMinfs to store information about which blocks
have been queued for allocation (in the
DataAllocationState), and an AllocateData method
which performs and commits the swap.

Additionally, start using methods to get/set the
size and block count of an inode, since in the
future this may be different for directory and data
vnodes.

Test: fs-test

Change-Id: I352ef8fb983c6b2ee53e7f463d73d31909ac0133