[inspect] Support writing on a VMO behind the facade.
This change creates new types:

Tree represents a single VMO tree, including methods to retrieve a
reference to the root object of the tree and the underlying VMO.

Inspector is an entrypoint into inspect to be used for creating Trees. A
future change will add multi-tree management functionality to the
Inspector.

CF-217: #progress

Change-Id: I7f92e1495f06c055d742bcf05d45231393d62fe9