[media][tests][e2e] Collapse VAD class hierarchy
To address code review feedback and make a cleaner implementation,
this CL moves consolidates ConfigImpl, DeviceImpl, InputImpl and
OutputImpl into a unified DeviceImpl. The Input and Output protocols
are composed of Config and Device, but to have a single implementation
with no forwarders between child and parent classes, it is cleaner to
put the implementation for both protocols into the one class.
Although it is an important part of our e2e testing efforts, the VAD
is test-only; this CL contains no product changes.

Test: build, CQ, overnight loops on various QEMU flavors (debug and
release) and NUC.

Change-Id: I5953cdcf9126550166bd21be51dc0a79f9f3bdc8