[inspect] Fix bugs with reader use after free.
We keep a stack of pairs of (parent*, index) to drive attaching children
to parents, however, the pointer is wrapped by a unique_ptr that has
values moved out of it.

This change replaces the pointer element in the pair with a block index
itself, and we add a new unordered map to map these block indices to the
parent ObjectHierarchy that is already in its final location (hence the
pointer is stable).

TEST=fx run-test inspect_unittest

Change-Id: Id577e5e560467266be65c1b0c5d4bfe3e910f821