[utest/async-loop] Fix waiting for exception packet to be processed
Previously we might call RunUntilIdle in between when a thread enters
the "in-exception" state and when the exception packet gets sent. It would
immediately return causing test failures. Instead wait for the packet
to be received.

All of this will need to be rewritten when exceptions-as-separate-objects
lands (where, for example, we might wait for the channel to be readable).
So instead of getting fancy, KISS.

FLK-78 #done

Change-Id: I322b0f9a131f9b4b1823c176b898e6f8edb1d476