[omaha-client] Initial check-in
This is the starting of the Omaha client service.  It's comprised of a
Rust crate which implements the base objects and protocol for
communicating with the Omaha service, and a service that uses that crate
to implement an Omaha client.

These packages are currently not part of the base build, and need to be
explicitly added to the build using 'fx set ... --available'.

Bug: PKG-448 #initial checkin
Test: Compiles and empty test binary executed with fx run-test
Change-Id: I890cb81fcbdf71f1f0d57710b96df58f8ad16da7