[debugger] Retrieving stdout/err from inferior.
This creates a socket that is passed upon creation to the inferior to be
used as the stdout/stderr file descriptors.

For now the agent only logs the streams. An integration test and sending
the data over to the client will be made in another CL.

TEST=Manual:

1. Process that writes to both stdout and stderr.
2. Having 2 process attached at the same time writing to stdout.

DX-595 # Comment.
DX-596 # Comment.
DX-724 # Comment.

Change-Id: If43b49c6de457f70ee8029a5f63abacf66a5c5af