[hikey][scripts] Pass full bootimage to hikey.
On Hikey, we place the Zircon.zbi image at 0x90000.

The bootloader also accepts a second argument for an initrd. If an
initrd is not provided, the hikey bootloader appears to hang for
30 seconds or so. As a workaround, we used to pass a dummy initrd
containing "foo".

The Hikey bootloader appears to place this bootimage 16MiB after
the kernel. Recently our arm64.zbi image has grown beyond 16MiB
(since it's a combined bootdata + kernel).

As a result the hikey bootloader was placing the dummy initrd
right in the middle of our arm64.zbi causing a corrupt bootdata
coming into userboot.

As a workaround, we pass the arm64.zbi as the initrd as well. When
the boot shim starts, it probes the initrd address and uses that zbi
instead of the one hanging off the end of the kernel.

Change-Id: Ic3c3c9c08ec896488ef135aa69cd5c6205eae902