[debugger] Launch components seamlessly.
Before this change, to the client launching a component would be indistinguishable
from a new process being caught from a filter. Furthermore, the agent
had no way of detecting that a particular process was being caught by
the ephemeral filters created for launching a component.

This change makes the agent track which filters are used for launching
components and associate an unique id to them. This ID is used by the
client to treat these process started events as a launched components
instead of a newly unexpected process.

This feature is not complete yet:

- Test the client in TargetImplTest that this feature works.
- Create a unique filter for each launch (currently ambiguous cases can
occur). Mask this unique name/filter from the client.
- Detect when launching a component fails and remove the state, both
from the agent and the client.
- Create an integration test for this case.

TEST=manual. Still under experimental flag.

DX-372  # Comment

Change-Id: I9807803307abfc46dc8d4af6e49af67d808b516c