[bt][tests] Add le.Central integration tests
Adds integration tests for exercising the Bluetooth LowEnergy Central
interface. This is done via a hermetic copy of bt-gap, which talks to a
host driver bound to a fake hci driver. The Tests themselves are run as
a component.

Test cases are written using asynchronous expectations on the Central
State, using a generic harness in the fuchsia-bluetooth library

Currently we have tests for exercising device scanning control.

Also adds a test for the fuchsia.bluetooth.control.Control interface.

NET-868 #done

Test: You can run the integration binary by running:
  $ fx run-test bluetooth-tests -t bt-host-integration-tests

Change-Id: I1ee1d0f8e5ce26360491b4fd36355b6ec8ead1aa