[debugger] Parse type names.
Parses type names and generates TypeExprNode objects. The symbol system
is not yet hooked up to identify type names currently, this will be
added in a later pass. Currently the tests use a mocked symbol query
function.

This is not currently hooked up to anything. It will be used in the
future to support casting.

Adds simple unit tests for parsing "true" and "false".

Change-Id: I69b888e4617aa04d66d504575b726dccd6305b3a