[bt][bt-gap] Don't halt bt-gap if we fail to start or stop a single host
Changes bt-gap so that if an error occurs when adding or removing a
host, this does not terminate the whole program. Due to timing issues,
it is possible during add_adapter that a bt-host device has been removed
before we attempt to initialize it (open a channel, set connectable,
enable background scan etc.), which results in an error. We should log
this error, but as this is a situation which could occur in normal
behavior, we should not terminate the program.

Test: This behavior can be seen by adding and quickly removing an hci
device before bt-gap has finished initializing it. This is hard to do as
a human but can be implemented in a test. This was previously uncovered
by the set_active_host test in bt-integration-tests, but the current
behavior no longer produces this.

Change-Id: Ife27b14dc6b711c9a4b7c9e3cebc5aa21ea742e3