[debugger] Add builtin base types.
Add a mapping for the builtin C/C++ types and the common stdint types.

This is currently unused, it will be used in the future as a fallback
for type lookup when a type name is not found in the debugged program.

This currently handles one-word types only. In a future pass we'll
handle things like "short signed int".

Converts BaseType's constructor's "size" parameter to uint32_t which
matches the type of the size use by the Type class.

Change-Id: I2fd49c0a4f6967f68f6553d7564d02e492ff8432