[sysmem] Support ram domain buffers
Some clients need to negotiate what level of cache coherency that is
used with a buffer. For example, the Intel display hardware isn't
cache-coherent with anything else, while the GPU is cache coherent with
the CPU, and can be configured to disable caching for some textures.

It's much more efficient to only require that data be in RAM in these
cases, rather than require extra cache flushing. Add a flag so that
clients can advertise RAM domain support, and some simple heuristics for
when to use it.

CPU coherency is still the default; advertising support for CPU
coherency is temporarily added to the display drivers, even though it's
not correct, until consumers that require it are moved off.

Change-Id: Ie4a7295d6c08223193a1c2e04a399a91112e34a9