[zircon][power] Add ZX_PROTOCOL_POWER
This changeset adds ZX_PROTOCOL_POWER that operates on a particular
power domain. This has the same semantics of GPIO for now. A subsystem will
request for a power domain in its board driver. The device driver will
get the power protocol on a power domain, using the index of that domain
in the array of power domains requested in their board driver.

Test: Used the power protocol apis in mt8167-touch and was able to
enable/ read status of a domain correctly.

Change-Id: I47a8809264e89e3abe466956baf079ac53dbdb90