[magma] Start support for non-cpu-domain sysmem
Struct members are added to specify whether memory should be allocated
in the RAM domain, and a function is added to check what coherency
domain the memory is in.

At the moment the new struct members are ignored. That will be changed
in a follow-up patch, after drivers are changed to initialize them.

Change-Id: I34f9a7215c8d4881baac0f15baeb6535478a2c04