[bt][tests] Rename bt-host-integration-tests, add README.md
What was previously bt-host-integration-tests no longer exercises solely
the bt-host subsystem, so the name is now a misnomer. We rename the
binary to bt-integration-tests to more accurately describe it. In
future, we may split the tests into multiple binaries if there is a need
for it.

Also add a README.md for the integration tests and link it from the top
level bluetooth README.md, to fill in a documentation gap.

Change-Id: Id4e3db89b37034ba39c75a056b3e980703c9a734