[gn] Add top-level BUILD.gn and forwarding targets to //build/gn:*
This adds a top-level BUILD.gn file and defines it as the root (by
removing the root_build override in //.gn) and adds forwarding rules to
the top-level target definitions in //build/gn/BUILD.gn. Also moves the
file generation steps to the top level BUILD.gn file.

The other rules in //build/gn/BUILD.gn should move too, but they are
currently referred to by infrastructure configuration files. They can
move after this lands and rolls to GI and the infra configs are updated.

Change-Id: I2043cc507944158cfe691db884d9fc79c2e80cc4