[rust_crates] Add crates harfbuzz, freetype, and their dependencies
Full list of new crates:
- core-foundation       OSRB-160
- core-foundation-sys   OSRB-160
- core-graphics         OSRB-159
- core-text             OSRB-162
- foreign-types         OSRB-161
- foreign-types-shared  OSRB-161
- freetype              OSRB-148
- harfbuzz              OSRB-150
- harfbuzz-sys          OSRB-151

Also:
- Add tiny_mirrors replacements for cmake and pkg-config crates, which are
optional transitive dependencies of the harfbuzz and freetype crates, but which
we do not need or support in our build system.
- In //third_party/rust_crates/.cargo/config, add overrides to bypass building
the native harfbuzz and freetype libraries inside their wrapper crates, and
instead link to the versions of these libraries already built in our tree.

TEXT-60 #comment

Change-Id: Ieb9cfb08bcc1c909168a413f2e2af8d558728cb8