[benchmarking] Log command when |CHECK|ing command exit status
Log the attempted command again if we are about to fail a |CHECK| due to
a non-zero exit status.  While it's technically possible to find this
information from the "Running ..." line above, there can be an arbitrary
amount of test logging in between that and the failure, making it
difficult to find when reading buildbot logs.

Test: CQ
Change-Id: If612934a42bf371a14d2c8239ed78d6a6a145b8c