[componentmgr] Use cm_rust native representation.
This change also makes related improvements:
- cm_rust adds a new type, CapabilityPath, for paths to make it easy to
get dirname/basename
- Tweak error classes so that cm_fidl_validator returns an error::Error
and cm_rust defines its own Error type
- cm_rust now represents `program` and `facets` as fidl (but keeps the
fidl_into converter). The reason being that componentmgr itself doesn't
consume `program`, it just needs to pass it along to the runner. It's
less clear at this point what should happen with `facets`, so I'm
making it a fidl for now to be consistent with `program`.

Change-Id: I886a5529bb74c98ed686062c53e27c222f74d4e4