[debugger] Parse C-style casts.
Adds C-style cast support to the parser.

Since symbol lookup is not yet hooked up to the parser, this is not
currently usable. It will be hooked up in a follow-up.

This leaves some refactoring needs for the existing casting code: (a)
the existing reinterpret_cast support (which is special-cased as a
function call) should generate the new CastExprNode, and (b) the
existing "Coerce" function should be folded into the new general-purpose
casting function. These changes are noted in the code.

Change-Id: I62ac8753ef25886497bdfbe300f15c195a8b0489