[debugger] Move ExprTokenType to a separate file.
The ExprToken needs to have more information associated with it to
support future extensions. As a first step, this moves the enum out of
the ExprToken class and into its own toplevel enum class.

Adds a helper function to ExprParser to convert a token to a parser
dispatch info record.

There should be no functionality change, this is mostly
search-and-replace.

There are some collateral chang-format updates.

Change-Id: I549aec08117ae3a025a0fec1a19ae6eaba7e2ac6