[auth] Persist `/data/accounts/list.json` for use by AccountManager
This enables the AccountManager to reload the account list upon
instantiation (i.e. power cycles). Whereas individual accounts are
already persisted for use with the AccountManager, the account manager
will not load them unless they are present in this list.

This file is written and read by AccountManager only, and is currently
stored in "deprecated-global-persistent-storage"
("isolated-persistent-storage" is blocked by account_parent_dir creation
being managed by the AccountManager, and not AccountHandler).

AUTH-106
Tested: CQ

Change-Id: Ibaeeffb616f827640e1b287e802ca91d5678d561