[bootsvc] Create new Items service
Create fuchsia.boot.Items to propagate ZBI items to other parts of the
system from bootsvc.

Rather than pass the ZBI VMO directly to subsequent processes, such as
the devcoordinator and the platform bus, we allow restricted access to
parts of the ZBI through the service directly to the platform bus.

ZX-3385
ZX-3654 #done
ZX-3659

Change-Id: I8cddbde929c764dda26f15190f2cbf245cf77784