[ddk] Add CLOCK_IMPL protocol.
Currently the same clock protocol is used by both the clients and the
clock implementation drivers. But the index argument to the Enable and Disable
calls do not have the same meaning in these two use cases.
For the client, the index is zero based, relative to the number of clocks
assigned to the driver, while for the implementation drivers, the index is
an absolute clock ID.

The platform bus driver currently maps the relative indices to the absolute
clock IDs. However this mapping will soon move to a generic clock driver that
will make this mapping based on metadata provided by the board driver instead.
To facilitate creating this new generic clock driver, this change splits the
clock protocols into two: ZX_PROTOCOL_CLOCK_IMPL will now be implemented by
the clock implementation drivers, while the clients will continue to use
ZX_PROTOCOL_CLOCK to access their clocks.

Bug: ZX-3746

TEST: Boot astro into full UI

Change-Id: I6e1fe36c61a79cfa4d9e3b242aabc426c8e08a06