[kernel][mp] temporarily comment out trying to clean up the per cpu unplug thread
At the moment there's a race with the per cpu unplug thread as it exits
It's possible for the unplug thread to signal that it's finished and the
creator thread to come along and unmap the stack while the unplug thread
is still running.

The better solution is a bit more complicated, so for now just comment
out the thread cleanup and print a warning.

ZX-2232 #workaround

Tested: built x86 and arm through shutdown cycles to make sure the race isn't hit.
Change-Id: I7c864556a824cb8b4070069e7bb23938b3c85047