[handle] add KernelHandle class
RAII wrapper for Dispatchers owned by the kernel which can be upgraded
to a full HandleOwner, otherwise will properly tear down the Dispatcher
when it goes out of scope.

The purpose here is to avoid tricky corner cases where the calling code
would otherwise have to manually call on_zero_handles(), particularly in
OOM situations when Handle creation fails.

The next step will be to change all Dispatcher::Create() methods to
return a KernelHandle instead of a RefPtr<Dispatcher>.

ZX-3612

Test: k ut handle
Change-Id: I3bf3f2bcf02f4987d27afa79ea533be5e18371ba