[fuchsia-component] Fix opening directory
This change allows the root directory of a ServiceFs to be opened. First
it intercepts open requests for "." and binds the channel to the root
node. Second it responds properly to GetAttr requests so that listing
the directory can succeed.

The implementation is still missing actually returning a list of child
nodes.

Change-Id: I885c99ca0fe68e943497f603dbf12a4d58fa92e7