[debugger] Move casting inside one function.
Collapses the existing Coerse function ane the ReintepretCast function
to the general CastExprValue function that takes a casting method.

Renames the "coerce" function to be an implicit cast.

Adds a token for "reinterpret_cast" and changes the parsing of this from
a function call to an explicitly parsed cast. Add tests for this case.

Removes the old type_parser* files that were used by the old
reinterpret_cast function. Now the main parser handles types.

Change-Id: Id6e5b6d7405230157814c5ad294c7a55210899be