[fuchsia-component] do not assume the peer is still connected
serve_connection_at uses signal_peer on the channel sent with the
Directory::Open request. The peer may be disconnected if it has already
sent requests and does not intend to send more: in that case,
signal_peer will fail with Status::PEER_CLOSED. We should continue
processing the request in that case.

Change-Id: Ib2a3a810ab622a6a7a91003feec61b3f2fd21618