[mediasession] Distinguish local & remote sessions.
This allows media session publishers to identify their session as local
or remote. Remote sessions should not compete for "Active" status on device.

This introduces the need for access to the collection of media sessions
independent of the "Active" session event. That is provided here in the
form of delta events.

Since the protocol may now have clients that do not care about events,
a method has been added to confirm receipt of events. The service will
tolerate some amount of laziness on the client's part, but after an
amount of events without receipt the server will stop sending events
to the client until a receipt is received.

TEST: added tests for the paginated read, delta events, ack tolerance,
      and remote session's relationship with active session status
Change-Id: I6468caf8da7e362e8b75ceec3da61de5c0295e34