[kernel][x88] Prevent panic loop in efi crashlog
If the thread_lock is held when a panic happens then the efi crashlog saving will attempt to
switch address spaces, which requires acquiring the thread_lock, leading to another panic.

This provides a _locked variant of vmm_set_active_aspace and optionally usages it when storing
the crashlog.

ZX-3750 #done

Test: x86 NUC + QEMU
Change-Id: Ia1f9f95cd595b24e45d6b25b837ce0d8438dacc1