[zxcrypt] Unseal zxcrypt in paver/block-watcher
Now that we have a way to explicitly unseal zxcrypt volumes, we should
disable the self-unlocking behavior and do the work of unsealing said
volumes.  There's two places that we need to do this:

* the paver, when we format the /data partition (and potentially when we
  drop files into /data)
* devmgr's block-watcher, where we normally automount the /data
  partition.

Test plan: generated a new zedboot image, then used that to pave a
pixelbook.  Verified that the new system came up and that I could still
get an `fx shell` on it, which demonstrates that the ssh authorized_keys
got installed by the paver successfully, and that after boot the device
was able to unlock and mount /data.

Change-Id: Ib599b7d09b173dbef8222125afb3f60cb46b7d57