[sysmgr] Suppress log errors for "optional" services
Trace manager is a core service in services.config, but some product
configurations don't contain trace manager. Attempts to connect to the
trace manager result in logs from package loader and sysmgr, which
generate a lot of spam.

In general these logs are helpful, so we solve this by adding an
"optional_services" attribute to services.config.

US-565 #done

TESTED=Built with trace_manager-less configuration, and observed logs
with and without optional_services

Change-Id: If3de93ec9549a58af0ee4ba758453523d959ce21