[debugger] Make token lookup table-based.
Previously token lookup was a big hardcoded switch statement which made
expanding the list of tokens annoying.

Moves to a table-based lookup approach so we can add tokens without
changing the tokenizer code itself. This also gives us a place to store
language-specific token information in the future (some tokens might be
valid in some languages but not others).

There should be no behavior change.

Change-Id: I4627d9f3d36b145b4313513bf0f549d2aadf6392