[svc] remove services
This is really not a useful abstraction, and the interior mutability of
the thing makes it harmful to my eyes.

Removed the heartbeat example to avoid refactoring it, since it exhibits
a pattern now considered harmful (events without flow control).

Refactored the echo example and compatibility test server to remove
gratuitous indirection.

Change-Id: I79428563dc4804a29aa66b21f479db6fb491f110