[media][tests][e2e] VAD: correct async dispatch
To address code review feedback, with this CL the control driver
stores the default async_dispatcher at the time the FIDL interface
request is received. Starting with the Enable and Disable, we begin
to use it to deliver completion callbacks on the primary dispatcher
thread (rather than arbitrary context, previously _assumed_ to be
the primary thread).

Although it is an important part of our e2e testing efforts, the VAD
is test-only; this CL contains no product changes.

Test: build, CQ, overnight loops on various QEMU flavors (debug and
release) and NUC.

Change-Id: I138ee2148c6f6072d781ecd43f5d5f3954fdd615