[debugger] Hook up basic types to casting.
This hooks up the builtin types (int, char*, etc.) to the type parsing
code such that things like "print (int*)foo" will work.

Main symbol lookup is not yet hooked up. That will be done in a future
pass.

Adds "void" to the list of builtin types. This required some additions
because of the weird way DWARF encodes these types (with no type
reference at all) is not compatible with us building up a type
hierarchy. This uses the "none" base type to represent void. Unit tests
are added for the handling of this.

Fixes C-style cast operator precedence. Previously "(void*)a[0]" would
be parsed as converting a to a void* and then taking the 0th item.

Change-Id: If8c3e1492241ac69f378dceed033d55fd184f744