[System Monitor] workaround negative slope in data
When asked to calculate the slope of the data the dockyard would assert against
regressive data to avoid returning a negative slope (it's not expected that this
would occur). Slight variation between getting the clock and the counters can
occur which can create a negative slope for graphed data. This CL works around
that by using a prior known-good value.

Test: none
Change-Id: I917d21f5491da36d168eb878f33bda7641985a87