[text] Fix invalid TextPoint creation past end in Ime
Previously, if a client tried to create a TextPoint with an offset that
put the point beyond the end of the text field, the character index
would be incorrectly set to char_to_byte.len() instead of
char_to_byte.len()-1, which resulted in a panic.

This change both adds a breaking test to TextFieldTestSuite so that
other implementations can test for similar bugs, as well as actually
fixes the bug itself.

Change-Id: Ia14ed43cacfc56fb5efa35f528c80afe8bdf1304