[persistent-storage] Fix partially broken allowlist mechanism
For real this time. fxr/262164 fixed it for almost everything but I
forgot to update the fuchsia_component template with the same fix.

Also renames the "whitelist" to an "allowlist", adds the log_listener
component that I missed adding to that list initially, and deletes other
stale (already migrated) entries from the list.

Change-Id: Ifa91d3ddbe0b6f05042bc7147dc77180af5dc660