[media][tests][e2e] Additional virtualaudio FIDL interfaces
This CL adds methods and events to fuchsia.virtualaudio.Configuration
and fuchsia.virtualaudio.Device, as needed to support the end-to-end
testing push. This includes notifications to the client upon SetGain,
SetFormat, GetBuffer, Start, Stop, and PositionNotify, as well as the
ability to retrieve current values for format, gain, buffer and
position.

This also adds a method to remove the added-by-default format range.

Test: build, CQ, new test cases,

Change-Id: I02df1e9d56eb5b7ba114a35701835950e4a4b4da