[media][tests] virtual_audio (CLI util) updates
This CL updates the CLI virtual_audio utility to include the recently-
added interfaces for retrieval and notification of format, gain, ring
buffer, and position, as well as notification of start and stop, plus
more flexibility in setting supported formats.

This utility is only installed as a part of the tests/media packages.
This CL contains no product changes.

Test: build, CQ, functionality verified on all archs and flavors

Change-Id: Ic69128098b6d1e7b0d66984b2f66bdec1a275188