[guest] Handle a PCI interrupt race condition
PCI interrupts should be asserted continuously until acknowledged and we
attempt to handle that by tracking pending interrupts. The best method we
have to do this in the base PCI class is to track interrupts that we
tried to send while interrupts were disabled and resend them when
interrupts are enabled again. This is racy in the sense that interrupts
may be disabled after we send the interrupt to the guest but before the
guest has handled the interrupt.

This patch removes that behavior and requires derived classes to define
an interrupt pending condition. In the case of VirtioPci this is the
IsrStatus register. This eliminates the race condition for virtio devices.

FLK-99 #done
FLK-88 #done
MAC-240 #done

TEST=guest_integration_tests

Change-Id: I2fa0c13591e2a4d5fc87835f625a5365f83dfeca